Update rubin_nights_service to include context feed service.
